# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: DNS Daddy management API
  version: '1.0'
  description: "The management API for a DNS Daddy protective DNS resolver.\n\nEvery self-hosted resolver\
    \ serves this specification from `/openapi.yaml`,\nso a control plane can be generated against the\
    \ exact build it is talking\nto rather than against documentation that may have drifted.\n\n## Authentication\n\
    \nAll `/api/v1/*` endpoints except `health`, `auth/login`, `auth/logout`, and\n`auth/session` require\
    \ one of:\n\n* `Authorization: Bearer dnsd_…` — an API token, for machine access.\n* A `dnsdaddy_session`\
    \ cookie, issued by `POST /api/v1/auth/login`.\n\nCreate API tokens from Settings in the dashboard,\
    \ or via `POST /api/v1/tokens`.\nA token's secret is returned exactly once, at creation.\n\n## Cross-origin\
    \ requests\n\nA request authenticated by **session cookie** that uses a method other than\n`GET`,\
    \ `HEAD`, or `OPTIONS` must carry a same-origin `Origin` or `Referer`\nheader, or it is rejected with\
    \ `403`. This is CSRF protection: the session\ncookie is `SameSite=Lax`, which does not cover every\
    \ cross-site case.\n\nRequests authenticated by **bearer token** are exempt — a browser never\nattaches\
    \ an `Authorization` header cross-site, so CSRF does not apply.\nProgrammatic clients should use an\
    \ API token and can ignore this entirely.\n\n## Resolver endpoints (not part of this API)\n\n* `POST|GET\
...
  license:
...
servers:
...
security:
- bearerAuth: []
- sessionCookie: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - Administration
      summary: Exchange the admin password for a session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - password
              properties:
                password:
                  type: string
                  format: password
                  writeOnly: true
      responses:
        '200':
          description: Signed in; a `dnsdaddy_session` cookie is set
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many failed attempts from this address
      operationId: postAuthLogin
...
  /api/v1/auth/password:
    post:
      tags:
      - Administration
      summary: Change the admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - currentPassword
              - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                  writeOnly: true
                newPassword:
                  type: string
                  format: password
                  minLength: 12
                  writeOnly: true
      responses:
        '200':
          description: Password updated
        '400':
          $ref: '#/components/responses/BadRequest'
      operationId: postAuthPassword
...
  /api/v1/activity/queries:
    get:
      tags:
      - Metrics
      summary: Hourly query and block counts
      parameters:
      - name: hours
        in: query
        schema:
          type: integer
          default: 24
          minimum: 1
          maximum: 720
      responses:
        '200':
          description: Hourly buckets, including hours with no traffic
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityBucket'
      operationId: getActivityQueries
  /api/v1/threats/categories:
    get:
      tags:
      - Metrics
      summary: Blocked counts grouped by category
      parameters:
      - name: hours
        in: query
        schema:
          type: integer
          default: 24
      responses:
        '200':
          description: Category breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryCount'
      operationId: getThreatsCategories
  /api/v1/threats/top-domains:
    get:
      tags:
      - Metrics
      summary: Most frequently blocked domains
      parameters:
      - name: days
        in: query
        schema:
          type: integer
          default: 7
      - name: limit
        in: query
        schema:
          type: integer
          default: 10
          maximum: 200
      responses:
        '200':
          description: Ranked domains
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopBlockedDomain'
      operationId: getThreatsTopDomains
  /api/v1/queries:
    get:
      tags:
      - Metrics
      summary: Query log, newest first
      description: 'Cursor-paginated. Pass the returned `nextCursor` to fetch the next page;

        a `nextCursor` of `0` means the result set is exhausted.



        For repeatable continuation use identical absolute since/until bounds. Invalid cursors return
        400 instead of restarting. Export cursors are separate.'
      parameters:
      - name: networkId
        in: query
        schema:
          type: string
      - name: action
        in: query
        schema:
          type: string
          enum:
          - allowed
          - blocked
          - error
      - name: category
        in: query
        schema:
          type: string
      - name: domain
        in: query
        schema:
          type: string
        description: Substring match
      - name: exactDomain
        in: query
        schema:
          type: string
        description: Exact normalised domain; intended for investigation continuation.
      - name: clientIp
        in: query
        schema:
          type: string
        description: Client IP; IPv4-mapped forms normalise to IPv4.
      - name: client
        in: query
        schema:
          type: string
        description: Alias for clientIp; conflicting aliases are rejected.
      - name: hours
        in: query
        schema:
          type: integer
          minimum: 0
          maximum: 8760
        description: 0 means all retained data; use hours or since, not both.
      - name: since
        in: query
        schema:
          type: string
          format: date-time
        description: Inclusive RFC 3339 lower bound, at stored millisecond precision; replaces the default
          hours.
      - name: until
        in: query
        schema:
          type: string
          format: date-time
        description: Inclusive RFC 3339 upper bound; exports default to capture time.
      - name: cursor
        in: query
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        schema:
          type: integer
          default: 100
          minimum: 1
          maximum: 500
      responses:
        '200':
          description: Query-log page
          content:
            application/json:
              schema:
                type: object
                properties:
                  queries:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryEvent'
                  nextCursor:
                    type: integer
                    format: int64
                  count:
                    type: integer
                  limit:
                    type: integer
        '400':
          description: Invalid cursor, address or time window.
      operationId: getQueries
...
  /api/v1/networks:
    get:
      tags:
      - Networks
      summary: List networks with 24-hour activity
      responses:
        '200':
          description: Networks
          content:
            application/json:
              schema:
                type: object
                properties:
                  networks:
                    type: array
                    items:
                      $ref: '#/components/schemas/NetworkWithActivity'
                  clientAccess:
                    $ref: '#/components/schemas/ClientAccess'
      operationId: getNetworks
    post:
      tags:
      - Networks
      summary: Create a network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/PublicAckRequired'
      operationId: postNetworks
  /api/v1/networks/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Networks
      summary: Get a network
      responses:
        '200':
          description: Network
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getNetworksById
    patch:
      tags:
      - Networks
      summary: Update a network
      description: Omitted fields are left unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PublicAckRequired'
      operationId: patchNetworksById
    delete:
      tags:
      - Networks
      summary: Delete a network
      description: The last remaining network cannot be deleted, and neither can the built-in Default
        network (`n_default`) — it is the policy applied to every unmatched client and the ad-hoc resolver
        access switch at once, and nothing recreates it.
      responses:
        '204':
          description: 'Deleted, and the effective client ACL was republished from what remains. That
            does not necessarily mean the network''s addresses are now refused: the ACL is a union with
            no deny rules, so a range the bootstrap list or another permitted network also covers goes
            on being served. `GET /api/v1/networks` reports what is actually admitted.'
        '200':
          description: Deleted, but the resolver could not re-read its configuration, so whether the network's
            addresses are still being served could not be confirmed. The condition is also reported by
            `GET /api/v1/diagnostics`, by `GET /api/v1/health` as `clientAclStale`, and by `dnsdaddy doctor`
            (which reads it from health, since the state lives in the running daemon's memory) — until
            a reload succeeds.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                    - deleted
                  warning:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'Refused permanently: this is the last remaining network, or the built-in Default
            network. The request is well-formed and resending it will not succeed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: deleteNetworksById
  /api/v1/networks/{id}/rotate-token:
    post:
      tags:
      - Networks
      summary: Issue a new DoH token, invalidating the old one
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Network with its new token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
      operationId: postNetworksByIdRotateToken
  /api/v1/policies:
    get:
      tags:
      - Policies
      summary: List policies
      responses:
        '200':
          description: Policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Policy'
      operationId: getPolicies
    post:
      tags:
      - Policies
      summary: Create a policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
      operationId: postPolicies
...
  /api/v1/policies/{id}/rules:
    post:
      tags:
      - Policies
      summary: Add one allow or block entry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - kind
              - domain
              properties:
                kind:
                  type: string
                  enum:
                  - allow
                  - block
                domain:
                  type: string
                  example: supplier.example.com
                note:
                  type: string
                  description: Free text, e.g. a ticket reference
      responses:
        '200':
          description: The updated policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
      operationId: postPoliciesByIdRules
  /api/v1/policies/{id}/rules/{kind}/{domain}:
    delete:
      tags:
      - Policies
      summary: Remove one allow or block entry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: kind
        in: path
        required: true
        schema:
          type: string
          enum:
          - allow
          - block
      - name: domain
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Removed
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deletePoliciesByIdRulesByKindByDomain
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API token, which always begins `dnsd_`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: dnsdaddy_session
...